+++ src/main/java/net/soulsweaponry/datagen/recipe/WeaponRecipes.java
@@ -436,6 +436,16 @@ public class WeaponRecipes {
                 .criterion("has_item", conditionsFromItemPredicates(ItemPredicate.Builder.create()
                         .tag(ModTags.Items.LORD_SOUL).build()))
                 .offerTo(consumer);
+        ShapedRecipeJsonBuilder.create(RecipeCategory.COMBAT, WeaponRegistry.MEHRUNES_RAZOR)
+                .input('S', ItemRegistry.SHARD_OF_UNCERTAINTY)
+                .input('N', Items.NETHERITE_INGOT)
+                .input('O', ModTags.Items.LORD_SOUL)
+                .pattern("N")
+                .pattern("O")
+                .pattern("S")
+                .criterion("has_item", conditionsFromItemPredicates(ItemPredicate.Builder.create()
+                        .tag(ModTags.Items.LORD_SOUL).build()))
+                .offerTo(consumer);
 
         WeaponRecipeProvider.smithingRecipeLordSoulCombat(Ingredient.ofItems(WeaponRegistry.BLOODTHIRSTER), WeaponRegistry.DARKIN_BLADE, consumer);
         WeaponRecipeProvider.smithingRecipeLordSoulCombat(Ingredient.ofItems(Items.GOLDEN_SWORD), WeaponRegistry.DAWNBREAKER, consumer);